requirements.txt
django-solo
requests
PyVirtualDisplay==0.2.1
selenium==3.11.0
django==1.11.13
django-registration==2.4.1
